# static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Currency Exchange API",
    "description": "API for managing currencies and exchange rates"
  },
  "paths": {
    "/currency": {
      "post": {
        "summary": "Create a new currency",
        "description": "",
        "responses": {
          "200": {
            "description": "Currency created successfully"
          },
          "400": {
            "description": "Currency already exists"
          }
        }
      },
      "get": {
        "summary": "Get all currencies",
        "description": "",
        "responses": {
          "200": {
            "description": "A list of all currencies"
          }
        }
      }
    },
    "/currency/{id}": {
      "get": {
        "summary": "Get a specific currency",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Currency details"
          },
          "404": {
            "description": "Currency not found"
          }
        }
      }
    },
    "/exchange_rate": {
      "post": {
        "summary": "Create a new exchange rate",
        "description": "",
        "responses": {
          "200": {
            "description": "Exchange rate created successfully"
          },
          "400": {
            "description": "Exchange rate already exists"
          }
        }
      },
      "get": {
        "summary": "Get all exchange rates",
        "description": "",
        "responses": {
          "200": {
            "description": "A list of all exchange rates"
          }
        }
      }
    },
    "/exchange_rate/{id}": {
      "get": {
        "summary": "Get a specific exchange rate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Exchange rate details"
          },
          "404": {
            "description": "Exchange rate not found"
          }
        }
      }
    },
    "/convert": {
      "post": {
        "summary": "Convert an amount from one currency to multiple other currencies",
        "description": "",
        "parameters": [
          {
            "name": "from",
            "in": "body",
            "description": "The ID of the currency to convert from",
            "required": true,
            "type": "integer"
          },
          {
            "name": "to",
            "in": "body",
            "description": "A list of IDs of the currencies to convert to",
            "required": true,
            "type": "array",
            "items": {
              "type": "integer"
            }
          },
          {
            "name": "amount",
            "in": "body",
            "description": "The amount of money to convert",
            "required": true,
            "type": "number",
            "format": "double"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of conversions",
            "schema": {
              "$ref": "#/definitions/Conversion"
            }
          },
          "400": {
            "description": "Bad request. User sent invalid data."
          },
          "404": {
            "description": "Not found. Requested resource could not be found."
          }
        }
      }
    }
  },
  "definitions": {
    "Conversion": {
      "type": "object",
      "properties": {
        "from": {
          "type": "string",
          "description": "The currency code of the original currency"
        },
        "to": {
          "type": "string",
          "description": "The currency code of the target currency"
        },
        "amount": {
          "type": "number",
          "format": "double",
          "description": "The converted amount in the target currency"
        }
      }
    }
  }
}